requirements.txt
PyMuPDF
sentence-transformers
faiss-cpu
pymongo
pytest
reportlab
langchain
